+++ packages/snap-networks-utils/jest.config.js
@@ -14,6 +14,11 @@ module.exports = merge(baseConfig, {
   // The display name when running multiple projects
   displayName,
 
+  coveragePathIgnorePatterns: [
+    ...(baseConfig.coveragePathIgnorePatterns ?? []),
+    '.*/__mocks__/',
+  ],
+
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {