+++ packages/stellar-wallet-snap/snap.config.ts
@@ -6,6 +6,9 @@ dotenv();
 
 const config: SnapConfig = {
   input: resolve(__dirname, 'src/index.ts'),
+  typescript: {
+    enabled: true,
+  },
   server: {
     port: 8080,
   },